# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API pour OC elites du congo",
        "description": "Oc elites du congo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://elite.com/api"
        }
    ],
    "paths": {
...
        "/admin/rencontres/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__RENCONTRES"
                ],
                "summary": "Ajouter une rencontre",
                "operationId": "667fc7f41c5bad02d84e6f21e92d2e7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "team_home_name": {
                                        "type": "string",
                                        "example": "Real madrid"
                                    },
                                    "team_away_name": {
                                        "type": "string",
                                        "example": "FC barcelone"
                                    },
                                    "team_home_logo": {
                                        "type": "file"
                                    },
                                    "team_away_logo": {
                                        "type": "file"
                                    },
                                    "team_home_score": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "team_away_score": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_rencontre": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "user": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/rencontres/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__RENCONTRES"
                ],
                "summary": "Renvoie les informations d'une rencontre par son ID",
                "operationId": "961d53fa18c672424d61040a8e272446",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/rencontres/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__RENCONTRES"
                ],
                "summary": "Edite les informations d'une rencontre",
                "operationId": "71eeeaec9980d97614721ec6ace7ae35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "team_home_name": {
                                        "type": "string",
                                        "example": "Real madrid"
                                    },
                                    "team_away_name": {
                                        "type": "string",
                                        "example": "FC barcelone"
                                    },
                                    "team_home_logo": {
                                        "type": "file"
                                    },
                                    "team_away_logo": {
                                        "type": "file"
                                    },
                                    "team_home_score": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "team_away_score": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_rencontre": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "user": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/rencontres/delete/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__RENCONTRES"
                ],
                "summary": "Supprime une rencontre par son ID",
                "operationId": "68f79124b853fc40413e6ff52875f55f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": [
            []
        ]
    }
}